requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
django-cors-headers==4.0.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
coreapi==2.3.3
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
